# docs/archive/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "World Weaver Memory Gateway API",
    "version": "1.0.0",
    "description": "MCP server providing access to a tripartite memory system (episodic, semantic, procedural) for Claude Code instances. Implements 17 tools for autobiographical events, knowledge graphs, and learned procedures with ACT-R activation, FSRS retrievability, and Hebbian learning.",
    "contact": {
      "name": "World Weaver Project",
      "url": "https://github.com/astoreyai/world-weaver"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local MCP server"
    }
  ],
  "paths": {
...
    "/tools/get_session_id": {
      "post": {
        "summary": "Get the current Claude Code instance session ID",
        "description": "Returns session identifier for this instance",
        "operationId": "get_session_id",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "field": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}